+++ dwave/cloud/client/base.py
@@ -73,8 +73,8 @@ from dwave.cloud.regions import get_regions, resolve_endpoints
 from dwave.cloud.upload import ChunkedData
 from dwave.cloud.events import dispatches_events
 from dwave.cloud.utils.http import PretimedHTTPAdapter, BaseUrlSession, default_user_agent
-from dwave.cloud.utils import (
-    datetime_to_timestamp, utcnow, cached, retried, is_caused_by)
+from dwave.cloud.utils.time import datetime_to_timestamp, utcnow
+from dwave.cloud.utils import cached, retried, is_caused_by
 
 __all__ = ['Client']
