+++ dwave/cloud/computation.py
@@ -36,8 +36,8 @@ from concurrent.futures import TimeoutError
 from dateutil.parser import parse
 from operator import itemgetter
 
-from dwave.cloud.utils import (
-    utcnow, datetime_to_timestamp, aliasdict, deprecated)
+from dwave.cloud.utils.time import utcnow, datetime_to_timestamp
+from dwave.cloud.utils import aliasdict, deprecated
 from dwave.cloud.exceptions import InvalidAPIResponseError
 
 # Use numpy if available for fast decoding
