+++ tests/test_utils.py
@@ -37,9 +37,10 @@ from dwave.cloud.utils.logging import (
 from dwave.cloud.utils.qubo import (
     uniform_iterator, uniform_get,
     active_qubits, generate_random_ising_problem)
+from dwave.cloud.utils.time import utcnow
 from dwave.cloud.utils import (
     NumpyEncoder, coerce_numpy_to_python,
-    default_text_input, utcnow, cached, retried, deprecated, aliasdict,
+    default_text_input, cached, retried, deprecated, aliasdict,
     hasinstance, exception_chain,
     is_caused_by, get_distribution, PackageNotFoundError, VersionNotFoundError)
